Fixing problem
Hey Sparanoid. I was working on these latest updates. I hit “Sync” in
GitHub desktop. It caused all of my commits to go to /sparanoid/slimack
on github.com.

I did a git reset —hard to revert to the last good commit. This push
request goes back to that commit. So sorry about this.